+++ lib/assets/javascripts/new-dashboard/components/Subscriptions/SubscriptionActions.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="u-flex" :class="{'u-flex__direction--column': mode === 'column', 'u-flex__align--center': mode !== 'column'}">
+  <div class="u-flex actions" :class="{'u-flex__direction--column': mode === 'column', 'u-flex__align--center': mode !== 'column'}">
     <div v-if="dataset.sync_status !== 'syncing'" class="u-flex u-flex__align--center"
       :class="{ disabled: dataset.status !== 'active' || dataset.sync_status !== 'synced' }">
       <SubscriptionButtonTooltip v-if="dataset.sync_status === 'unsynced' && (dataset.unsynced_errors && dataset.unsynced_errors.length > 0)">
@@ -54,10 +54,12 @@
       </span>
     </div>
     <div v-if="mode !== 'column'" class="white-separator u-ml--12 u-mr--12"></div>
-    <div :class="{ 'u-mt--12': mode === 'column', disabled: dataset.status !== 'active' }">
+    <div :class="{ 'u-mt--12': mode === 'column', disabled: dataset.status !== 'active', 'u-flex u-flex__align--center': true }">
       <a class="text is-caption" :class="smallClass" href="#" @click="downloadNotebook">
         Explore with CARTOFrames
       </a>
+      <span class="u-ml--8 u-mr--8">|</span>
+      <SubscriptionsQuickActions @onAccess="openAccess"></SubscriptionsQuickActions>
     </div>
   </div>
 </template>
@@ -66,11 +68,13 @@
 
 import { mapState } from 'vuex';
 import SubscriptionButtonTooltip from './SubscriptionButtonTooltip';
+import SubscriptionsQuickActions from './SubscriptionsQuickActions';
 
 export default {
   name: 'SubscriptionActions',
   components: {
-    SubscriptionButtonTooltip
+    SubscriptionButtonTooltip,
+    SubscriptionsQuickActions
   },
   props: {
     mode: {
@@ -96,6 +100,21 @@ export default {
     }
   },
   methods: {
+    openAccess (platform) {
+      this.$store.commit('catalog/setCurrentSubscription', this.dataset);
+      this.$store.commit('catalog/setCurrentAccessPlatform', platform);
+      this.sendMetrics(platform);
+    },
+    sendMetrics (platform) {
+      if (platform === 'bigquery') platform = 'bq';
+      this.$store.dispatch(
+        'catalog/sendAccessAttemptMetrics',
+        {
+          datasetId: this.dataset.id,
+          platform,
+          licenseType: this.dataset.license_type
+        });
+    },
     downloadNotebook (e) {
       e.preventDefault();
       this.$store.dispatch('catalog/downloadNotebook', { id: this.dataset.slug, type: this.dataset.type });
@@ -110,8 +129,12 @@ export default {
 
 <style scoped lang="scss">
 @import 'new-dashboard/styles/variables';
+.actions {
+  position: relative
+}
 .disabled {
-  >a {
+  >a,
+  >.quick-actions {
     opacity: 0.4;
     pointer-events: none;
   }